Use bigger integer type for num_updates field of a project
For the record, smallint value was overflowed. Top 5 projects are
nginx, llvm, spidermonkey, asterisk and mplayer. It may be that the
actual cause is flapping data in some repo.

+++ sql.d/schema/create_schema_tables.sql
@@ -84,7 +84,7 @@ CREATE TABLE metapackages (
 	num_families_newest smallint NOT NULL DEFAULT 0,
 	has_related boolean NOT NULL DEFAULT false,
 	has_cves boolean NOT NULL DEFAULT false,
-	num_updates smallint NOT NULL DEFAULT 0,
+	num_updates integer NOT NULL DEFAULT 0,
 
 	first_seen timestamp with time zone NOT NULL DEFAULT now(),
 	orphaned_at timestamp with time zone
